+++ app/components/settingsElement.js
