+++ app/routes/drawer/routes/settings/generalAppInfo/generalAppInfoScene.js
@@ -1,7 +1,7 @@
 import React, { Component, PropTypes } from 'react';
 import { View, StyleSheet, Dimensions } from 'react-native';
 import Header from '../../../../../components/header';
-import LabelSettings from '../../../../../components/labelSettings';
+import SettingsElement from '../../../../../components/settingsElement';
 const window = Dimensions.get('window');
 
 class generalAppInfoScene extends Component {
@@ -16,9 +16,9 @@ class generalAppInfoScene extends Component {
             <View style={{ flex: 1 }}>
                 <Header leftIcon="arrow-left" navigatorLeft={() => this.props.navigation.goBack()} title="General Info" />
                 <View style={styles.container}>
-                    <LabelSettings leftElement="Version:" rightElement="0.0.1" />
-                    <LabelSettings leftElement="Build:" rightElement="1" />
-                    <LabelSettings leftElement="Authors:" rightElement="ReactPoland" />
+                    <SettingsElement leftElement="Version:" rightElement="0.0.1" />
+                    <SettingsElement leftElement="Build:" rightElement="1" />
+                    <SettingsElement leftElement="Authors:" rightElement="ReactPoland" />
                 </View>
             </View>
         );
